requirements.txt
chromadb
sentence-transformers
python-dotenv
google-generativeai
pdfplumber
pymupdf
nltk
